Match code_BD670, code_B6580, code_B3A70, code_B2900 (#80)
* match: src/app/code_BD670

* match: src/app/code_B6580

* match: most of app/code_B2900

* refac: re-define Unk80364210, introduce IS_MAIN_VEHICLE and IS_NOT_MAIN_VEHICLE

* match: func_8032B560

* rematch cannonLandedFrame

* refac: drop useless typedef

* match: func_8032C540, func_8032CC44

* match: func_8032CF28, func_8032D33C

* match: app/code_B3A70.c

* renames and stuff

* rename rocket_pack -> rocket_belt

* review feedback

* Update src/app/code_B3A70.c

* Apply suggestions from code review

+++ src/app/code_72910.h
@@ -4,7 +4,7 @@
 #include <uv_level.h>
 
 void func_802EB3E0(void);
-void func_802EB424(Mtx4F*, s32);
+void func_802EB424(Mtx4F*, f32);
 void func_802EB598(void);
 void func_802EB5E4(void);
